Update most of remaining project naming
