+++ app/src/main/java/com/dpanger/vehicles/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.dpanger.recipes
+package com.dpanger.vehicles
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -6,13 +6,10 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.dpanger.recipes.home.HomeScreen
-import com.dpanger.recipes.themes.RecipesTheme
+import com.dpanger.vehicles.home.HomeScreen
+import com.dpanger.vehicles.themes.VehiclesTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -21,7 +18,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
 
         setContent {
-            RecipesTheme {
+            VehiclesTheme {
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
