+++ app/src/main/java/com/dpanger/vehicles/VehiclesApplication.kt
@@ -1,7 +1,7 @@
-package com.dpanger.recipes
+package com.dpanger.vehicles
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class RecipesApplication : Application()
+class VehiclesApplication : Application()
