+++ data-multiplatform/src/main/java/com/dpanger/vehicles/data/Recipe.kt
@@ -1,4 +1,4 @@
-package com.dpanger.recipes.data
+package com.dpanger.vehicles.data
 
 data class Recipe(
     val title: String
