+++ data/src/main/java/com/dpanger/vehicles/data/DataModule.kt
@@ -1,4 +1,4 @@
-package com.dpanger.recipes.data
+package com.dpanger.vehicles.data
 
 import dagger.Module
 import dagger.Provides
