+++ home/src/main/java/com/dpanger/vehicles/home/LoadRecipesUseCase.kt
@@ -1,6 +1,6 @@
-package com.dpanger.recipes.home
+package com.dpanger.vehicles.home
 
-import com.dpanger.recipes.data.RecipeRepository
+import com.dpanger.vehicles.data.RecipeRepository
 import kotlinx.collections.immutable.toImmutableList
 import javax.inject.Inject
